WIP: Fix some unit tests hanging indefinitely
Dynamics that have neural networks have a __syncthreads() call in them
at some point due to internal neural network connections. The
StepTestKernel used in the dynamics unit tests have if statements which
reduce the threads to the number of queries. If the number of x threads
doesn't evenly divide the number of queries, this would cause not all
threads to reach __syncthreads(). This commit makes sure that all the
offending unit tests are fixed to have even division.